Refactor: Move `processNextBlock` into own file
